feat(tui): add source-line-range windowing for the diff overlay
rows_in_source_range slices an overlay's row list down to the source
view's already-clamped scroll window, anchoring each Removed row
(which has no line number of its own) to the Unchanged/Added row it
precedes. Split tests into source_diff/tests.rs (file-size
discipline: production+test would otherwise cross 700 lines).
