+++ rinkaku-tui/src/source_diff.rs
@@ -6,6 +6,10 @@
 
 use crate::diff_view::{DiffLineKind, FileHunks, Hunk};
 
+#[cfg(test)]
+#[path = "source_diff/tests.rs"]
+mod tests;
+
 /// One diff-only line an overlay adds beyond what [`crate::source::SourceView`]
 /// already renders, either a new-side line already present in the source
 /// (`Added`) or a line that no longer exists there (`Removed`).
@@ -165,6 +169,69 @@ pub fn overlay_source_lines(
     Some(rows)
 }
 
+/// Slices `rows` (an [`overlay_source_lines`] result) down to the rows
+/// belonging to the 1-based inclusive source line range
+/// `[start_line, end_line]` — the source view's already-clamped scroll
+/// window (`crate::ui::source_screen::clamped_window`, which operates on
+/// raw source line indices, unaffected by how many extra `Removed` rows an
+/// overlay inserts).
+///
+/// A `Removed` row has no `line_number` of its own (`OverlayRow`'s own doc
+/// comment); its *anchor* is the `Unchanged`/`Added` row immediately
+/// following it in `rows`' own order (or, for a run of `Removed` rows at
+/// the very end of `rows` — a deletion at end of file, `overlay_source_lines`'s
+/// own trailing-position handling — the source's last line, since that run
+/// renders only once the reviewer has scrolled to see the file's end). A
+/// `Removed` row is included exactly when its anchor line number falls in
+/// `[start_line, end_line]`.
+pub fn rows_in_source_range(
+    rows: &[OverlayRow],
+    start_line: usize,
+    end_line: usize,
+) -> &[OverlayRow] {
+    let anchors = row_anchor_line_numbers(rows);
+
+    let first = anchors.iter().position(|anchor| *anchor >= start_line);
+    let Some(first) = first else {
+        return &[];
+    };
+    let last = anchors[first..]
+        .iter()
+        .position(|anchor| *anchor > end_line)
+        .map(|offset| first + offset)
+        .unwrap_or(rows.len());
+
+    &rows[first..last]
+}
+
+/// One entry per `rows`, the 1-based source line number that row is
+/// anchored to for range-membership purposes ([`rows_in_source_range`]'s
+/// own doc comment on what "anchor" means for a `Removed` row).
+fn row_anchor_line_numbers(rows: &[OverlayRow]) -> Vec<usize> {
+    let last_line_number =
+        rows.iter()
+            .filter_map(|row| match row {
+                OverlayRow::Unchanged { line_number, .. }
+                | OverlayRow::Added { line_number, .. } => Some(*line_number),
+                OverlayRow::Removed { .. } => None,
+            })
+            .next_back()
+            .unwrap_or(0);
+
+    let mut anchors = vec![0; rows.len()];
+    for (index, row) in rows.iter().enumerate().rev() {
+        anchors[index] = match row {
+            OverlayRow::Unchanged { line_number, .. } | OverlayRow::Added { line_number, .. } => {
+                *line_number
+            }
+            OverlayRow::Removed { .. } => {
+                anchors.get(index + 1).copied().unwrap_or(last_line_number)
+            }
+        };
+    }
+    anchors
+}
+
 /// Whether every `Context` line in `hunk` matches `source_lines` at the
 /// position [`hunk_overlay_lines`]'s own walk would place it — the drift
 /// check backing [`overlay_source_lines`]'s `None` return (ADR 0046
@@ -194,333 +261,4 @@ fn hunk_matches_source(hunk: &Hunk, source_lines: &[String]) -> bool {
         }
     }
     true
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::diff_view::DiffLine;
-    use pretty_assertions::assert_eq;
-
-    fn diff_line(kind: DiffLineKind, content: &str) -> DiffLine {
-        DiffLine {
-            kind,
-            content: content.to_string(),
-        }
-    }
-
-    // --- hunk_overlay_lines ---
-
-    #[test]
-    fn should_return_empty_when_hunk_has_no_new_range() {
-        let hunk = Hunk {
-            header: "@@ garbage @@".to_string(),
-            new_range: None,
-            lines: vec![diff_line(DiffLineKind::Context, "fn a() {}")],
-        };
-
-        let actual = hunk_overlay_lines(&hunk);
-
-        assert_eq!(Vec::<HunkOverlayLine>::new(), actual);
-    }
-
-    #[test]
-    fn should_position_added_lines_by_new_side_line_number_for_a_pure_addition() {
-        let hunk = Hunk {
-            header: "@@ -1,1 +1,3 @@".to_string(),
-            new_range: Some((1, 3)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Added, "fn b() {}"),
-                diff_line(DiffLineKind::Added, "fn c() {}"),
-            ],
-        };
-
-        let expected = vec![
-            HunkOverlayLine {
-                position: 2,
-                kind: OverlayLineKind::Added,
-                content: "fn b() {}".to_string(),
-            },
-            HunkOverlayLine {
-                position: 3,
-                kind: OverlayLineKind::Added,
-                content: "fn c() {}".to_string(),
-            },
-        ];
-        let actual = hunk_overlay_lines(&hunk);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_position_removed_lines_before_the_new_side_line_they_used_to_precede() {
-        let hunk = Hunk {
-            header: "@@ -1,3 +1,2 @@".to_string(),
-            new_range: Some((1, 2)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Removed, "fn old() {}"),
-                diff_line(DiffLineKind::Context, "fn c() {}"),
-            ],
-        };
-
-        let expected = vec![HunkOverlayLine {
-            position: 2,
-            kind: OverlayLineKind::Removed,
-            content: "fn old() {}".to_string(),
-        }];
-        let actual = hunk_overlay_lines(&hunk);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_position_removed_line_at_file_end_when_hunk_is_a_trailing_pure_deletion() {
-        let hunk = Hunk {
-            header: "@@ -3,1 +2,0 @@".to_string(),
-            new_range: Some((3, 2)),
-            lines: vec![diff_line(DiffLineKind::Removed, "fn tail() {}")],
-        };
-
-        let expected = vec![HunkOverlayLine {
-            position: 3,
-            kind: OverlayLineKind::Removed,
-            content: "fn tail() {}".to_string(),
-        }];
-        let actual = hunk_overlay_lines(&hunk);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_position_mixed_added_and_removed_lines_in_one_replace_hunk() {
-        let hunk = Hunk {
-            header: "@@ -1,3 +1,3 @@".to_string(),
-            new_range: Some((1, 3)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Removed, "fn old() {}"),
-                diff_line(DiffLineKind::Added, "fn new() {}"),
-                diff_line(DiffLineKind::Context, "fn c() {}"),
-            ],
-        };
-
-        let expected = vec![
-            HunkOverlayLine {
-                position: 2,
-                kind: OverlayLineKind::Removed,
-                content: "fn old() {}".to_string(),
-            },
-            HunkOverlayLine {
-                position: 2,
-                kind: OverlayLineKind::Added,
-                content: "fn new() {}".to_string(),
-            },
-        ];
-        let actual = hunk_overlay_lines(&hunk);
-
-        assert_eq!(expected, actual);
-    }
-
-    // --- overlay_source_lines ---
-
-    fn file_hunks(hunks: Vec<Hunk>) -> FileHunks {
-        FileHunks {
-            path: "src/lib.rs".to_string(),
-            hunks,
-        }
-    }
-
-    #[test]
-    fn should_return_empty_rows_when_file_has_no_hunks() {
-        let source_lines = vec!["fn a() {}".to_string()];
-        let hunks = file_hunks(vec![]);
-
-        let expected = vec![OverlayRow::Unchanged {
-            line_number: 1,
-            content: "fn a() {}".to_string(),
-        }];
-        let actual = overlay_source_lines(&source_lines, &hunks).expect("no drift to detect");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_tag_a_pure_addition_as_added_rows_at_their_source_positions() {
-        let source_lines = vec![
-            "fn a() {}".to_string(),
-            "fn b() {}".to_string(),
-            "fn c() {}".to_string(),
-        ];
-        let hunks = file_hunks(vec![Hunk {
-            header: "@@ -1,2 +1,3 @@".to_string(),
-            new_range: Some((1, 3)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Added, "fn b() {}"),
-                diff_line(DiffLineKind::Context, "fn c() {}"),
-            ],
-        }]);
-
-        let expected = vec![
-            OverlayRow::Unchanged {
-                line_number: 1,
-                content: "fn a() {}".to_string(),
-            },
-            OverlayRow::Added {
-                line_number: 2,
-                content: "fn b() {}".to_string(),
-            },
-            OverlayRow::Unchanged {
-                line_number: 3,
-                content: "fn c() {}".to_string(),
-            },
-        ];
-        let actual = overlay_source_lines(&source_lines, &hunks).expect("no drift to detect");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_insert_removed_row_immediately_before_its_new_side_position() {
-        let source_lines = vec!["fn a() {}".to_string(), "fn c() {}".to_string()];
-        let hunks = file_hunks(vec![Hunk {
-            header: "@@ -1,3 +1,2 @@".to_string(),
-            new_range: Some((1, 2)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Removed, "fn old() {}"),
-                diff_line(DiffLineKind::Context, "fn c() {}"),
-            ],
-        }]);
-
-        let expected = vec![
-            OverlayRow::Unchanged {
-                line_number: 1,
-                content: "fn a() {}".to_string(),
-            },
-            OverlayRow::Removed {
-                content: "fn old() {}".to_string(),
-            },
-            OverlayRow::Unchanged {
-                line_number: 2,
-                content: "fn c() {}".to_string(),
-            },
-        ];
-        let actual = overlay_source_lines(&source_lines, &hunks).expect("no drift to detect");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_append_removed_rows_at_the_tail_when_deletion_is_at_file_end() {
-        let source_lines = vec!["fn a() {}".to_string(), "fn b() {}".to_string()];
-        let hunks = file_hunks(vec![Hunk {
-            header: "@@ -1,3 +1,2 @@".to_string(),
-            new_range: Some((3, 2)),
-            lines: vec![diff_line(DiffLineKind::Removed, "fn tail() {}")],
-        }]);
-
-        let expected = vec![
-            OverlayRow::Unchanged {
-                line_number: 1,
-                content: "fn a() {}".to_string(),
-            },
-            OverlayRow::Unchanged {
-                line_number: 2,
-                content: "fn b() {}".to_string(),
-            },
-            OverlayRow::Removed {
-                content: "fn tail() {}".to_string(),
-            },
-        ];
-        let actual = overlay_source_lines(&source_lines, &hunks).expect("no drift to detect");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_composite_multiple_hunks_across_the_same_file() {
-        let source_lines = vec![
-            "fn a() {}".to_string(),
-            "fn b() {}".to_string(),
-            "fn x() {}".to_string(),
-            "fn y() {}".to_string(),
-        ];
-        let hunks = file_hunks(vec![
-            Hunk {
-                header: "@@ -1,1 +1,2 @@".to_string(),
-                new_range: Some((1, 2)),
-                lines: vec![
-                    diff_line(DiffLineKind::Context, "fn a() {}"),
-                    diff_line(DiffLineKind::Added, "fn b() {}"),
-                ],
-            },
-            Hunk {
-                header: "@@ -8,1 +9,2 @@".to_string(),
-                new_range: Some((3, 4)),
-                lines: vec![
-                    diff_line(DiffLineKind::Context, "fn x() {}"),
-                    diff_line(DiffLineKind::Added, "fn y() {}"),
-                ],
-            },
-        ]);
-
-        let expected = vec![
-            OverlayRow::Unchanged {
-                line_number: 1,
-                content: "fn a() {}".to_string(),
-            },
-            OverlayRow::Added {
-                line_number: 2,
-                content: "fn b() {}".to_string(),
-            },
-            OverlayRow::Unchanged {
-                line_number: 3,
-                content: "fn x() {}".to_string(),
-            },
-            OverlayRow::Added {
-                line_number: 4,
-                content: "fn y() {}".to_string(),
-            },
-        ];
-        let actual = overlay_source_lines(&source_lines, &hunks).expect("no drift to detect");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_return_none_when_context_line_does_not_match_source_at_its_position() {
-        let source_lines = vec!["fn a_edited() {}".to_string(), "fn c() {}".to_string()];
-        let hunks = file_hunks(vec![Hunk {
-            header: "@@ -1,2 +1,2 @@".to_string(),
-            new_range: Some((1, 2)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Context, "fn c() {}"),
-            ],
-        }]);
-
-        let actual = overlay_source_lines(&source_lines, &hunks);
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_return_none_when_context_line_position_is_past_the_end_of_source() {
-        let source_lines = vec!["fn a() {}".to_string()];
-        let hunks = file_hunks(vec![Hunk {
-            header: "@@ -1,1 +1,2 @@".to_string(),
-            new_range: Some((1, 2)),
-            lines: vec![
-                diff_line(DiffLineKind::Context, "fn a() {}"),
-                diff_line(DiffLineKind::Context, "fn missing() {}"),
-            ],
-        }]);
-
-        let actual = overlay_source_lines(&source_lines, &hunks);
-
-        assert_eq!(None, actual);
-    }
 }
